Add initial project structure
This patch adds the initial project structure, including licenses,
linters and their configuration, and a basic test case that ensures that
the module can be loaded.

Most tooling is taken over from pynitrokey.  One notable exception is
that flit is replaced by poetry.  This only affects developers, not end
users.  We recently made the same change in nitrokey-app2 [0], so I
think it makes sense to use the same tooling here.

[0] https://github.com/Nitrokey/nitrokey-app2/pull/172

In the CI builds, we check that both the lockfile and the latest
dependency versions work.  It would good to also run it for the minimum
versions specified in pyproject.toml, but this is currently not
supported by poetry or pip:

https://github.com/python-poetry/poetry/issues/3527
https://github.com/pypa/pip/issues/8085