Rename Device to Model
Previously, we had the TrussedDevice abstract class implemented by the
device classes NK3 and NKPK, and the Device enum that lists the
supported devices.  To reduce the potential confusion between the actual
implementation and the enum, this patch renames Device to Model (and
DeviceData to ModelData) to make clear that it does not refer to a
specific instance.