nrf52_upload: Remove dfu_trigger
DFU Trigger is a mechanism for rebooting from firmware into bootloader
[0].  It is used if the device does not have one of the known bootloader
VID/PID combinations.  As the Nitrokey 3 bootloader is not part of the
whitelist, it is currently executed for every firmware update.  But the
Nitrokey 3 already is in bootloader mode and does not support the DFU
trigger protocol, so it just delays the firmware update by 5 seconds.

This patch removes the DFU trigger implementation.

[0] https://infocenter.nordicsemi.com/index.jsp?topic=%2Fcom.nordic.infocenter.sdk5.v15.3.0%2Flib_dfu_trigger_usb.html