ci: Remove full compatibility tests from default workflow
The full compatibility tests with all Python versions on all operating
systems lead to a significant slowdown when multiple workflows are
running.  With this patch, we only run tests with Python 3.9 on the
three operating systems.  The full tests are run once a day or after a
manual trigger (e. g. before a release).