nk3.updates: Replace update paths with set of migrations
Previously, we determined a single update path to trigger
version-specific behavior.  But it is possible that multiple update
paths apply to the same update if versions have been skipped.  This
patch replaces the single update path with a set of migrations to make
this more flexible.

As this is an implementation detail of the Updater, the relevant types
are removed from the public API.