nk3.updates: Run migration checks early if possible
Previously, we ran the migration checks after rebooting into bootloader
because we needed to hardware variant and there was no way to query it
from the firmware.  In the mean time, we have added the hardware variant
to the status reported by the admin app.  So if we have a more recent
firmware version (or if the device is already in bootloader mode), we
can execute the checks before rebooting into bootloader mode and fail
early if there is a problem.