nk3.updates: Remove reboot during update
In v0.3.0, we added a reboot during the update to ensure that the device
status is up to date.  This was by mistake – we decided to skip this
change for better compatibility with QubesOS.  The current implementation
was broken anyway as the device handle used for the reboot to bootloader
was invalidated by the first reboot.