ci: Test different dependency resolution strategies with uv
This patch adds a job to the CI that runs the type checks and the test
suite with uv using different dependency resolution strategies:  lowest
will select the lowest dependency versions that match our requirements,
lowest-direct will select the lowest version for direct dependencies and
the highest version for transitive dependencies, and highest will select
the highest version for all dependencies.

This helps to ensure that new code also works with old dependency
versions.